LPS-188018 Extract the client interface but don't expose it
